Update to use clippy 1.67 in CI (#810)

+++ src/gridfs/download.rs
@@ -192,7 +192,7 @@ impl GridFsBucket {
         if n != file.n() {
             return Err(ErrorKind::GridFs(GridFsErrorKind::WrongNumberOfChunks {
                 actual_number: n,
-                expected_number: file.n() as u32,
+                expected_number: file.n(),
             })
             .into());
         }
